refactor: DayCard.tsx aufbrechen in day-card/ Directory (#210)
1.330-Zeilen God-Component in 11 fokussierte Dateien zerlegt:
- DayCard.tsx (333Z) als Orchestrator
- SessionDetailDialog.tsx (441Z) mit SessionReadOnlyView
- RestDayDialog.tsx (155Z)
- DraggableSessionRow/DraggableRestDay (dnd-kit Wrapper)
- SessionCardRow, SegmentRow (UI-Komponenten)
- day-card-constants.ts, day-card-helpers.ts
- Barrel-Export via index.ts

RUN_TYPE_OPTIONS in constants/plan.ts konsolidiert (Single Source of Truth).

Visuell identisches Verhalten. TSC + ESLint + Vitest grün.
