+++ frontend/src/components/day-card/day-card-constants.ts
@@ -1,0 +1,25 @@
+/** DayCard-interne Konstanten. Shared Constants in @/constants/plan. */
+
+export const INITIAL_TYPE_OPTIONS = [
+  { value: '', label: 'Leer' },
+  { value: 'running', label: 'Laufen' },
+  { value: 'strength', label: 'Kraft' },
+  { value: 'rest', label: 'Ruhetag' },
+];
+
+export const MAX_SESSIONS = 3;
+
+export const TYPE_ICON_COLORS: Record<string, string> = {
+  easy: 'text-[var(--color-primary-1-500)]',
+  recovery: 'text-[var(--color-primary-1-400)]',
+  tempo: 'text-[var(--color-primary-2-500)]',
+  intervals: 'text-[var(--color-primary-2-600)]',
+  long_run: 'text-[var(--color-primary-1-600)]',
+  progression: 'text-[var(--color-primary-1-600)]',
+  repetitions: 'text-[var(--color-primary-2-600)]',
+  fartlek: 'text-[var(--color-primary-2-400)]',
+  race: 'text-[var(--color-primary-2-700)]',
+  strength: 'text-[var(--color-secondary-1-500)]',
+  rest: 'text-[var(--color-text-muted)]',
+  empty: 'text-[var(--color-text-disabled)]',
+};
