+++ frontend/src/components/day-card/day-card-helpers.ts
@@ -1,0 +1,29 @@
+import type { PlannedSession, WeeklyPlanEntry } from '@/api/weekly-plan';
+
+export function getSessionTypeKey(session: PlannedSession): string {
+  if (session.training_type === 'strength') return 'strength';
+  if (session.training_type === 'running' && session.run_details?.run_type) {
+    return session.run_details.run_type;
+  }
+  if (session.training_type === 'running') return 'easy';
+  return 'empty';
+}
+
+export function getEntryTypeKey(entry: WeeklyPlanEntry): string {
+  if (entry.is_rest_day) return 'rest';
+  if (entry.sessions.length === 0) return 'empty';
+  return getSessionTypeKey(entry.sessions[0]);
+}
+
+export function getDateStr(weekStart: string, dayOfWeek: number): string {
+  const d = new Date(weekStart);
+  d.setDate(d.getDate() + dayOfWeek);
+  return d.getDate().toString();
+}
+
+export function isDayInPast(weekStart: string, dayOfWeek: number): boolean {
+  const d = new Date(weekStart);
+  d.setDate(d.getDate() + dayOfWeek);
+  d.setHours(23, 59, 59, 999);
+  return d < new Date();
+}
