+++ frontend/src/components/day-card/index.ts
@@ -1,0 +1,1 @@
+export { DayCard } from './DayCard';
